refactor(store): extract discovery store from home, keep home behavior intact
